+++ tests/TestCase/src/Controller/BlogPostsControllerTest.php
@@ -15,9 +15,12 @@
  */
 use App\Test\TestCase\AppCakeTestCase;
 use Cake\ORM\TableRegistry;
+use Cake\TestSuite\IntegrationTestTrait;
 
 class BlogPostsControllerTest extends AppCakeTestCase
 {
+
+    use IntegrationTestTrait;
 
     public $BlogPost;
 
@@ -29,25 +32,26 @@ class BlogPostsControllerTest extends AppCakeTestCase
 
     public function testBlogPostDetailOnlinePublicLoggedOut()
     {
-        $this->httpClient->get($this->Slug->getBlogPostDetail(2, 'Demo Blog Artikel'));
-        $this->assert200OkHeader();
+        $this->get($this->Slug->getBlogPostDetail(2, 'Demo Blog Artikel'));
+        $this->assertResponseCode(200);
     }
 
     public function testBlogPostDetailOfflinePublicLoggedOut()
     {
         $blogPostId = 2;
         $this->changeBlogPost($blogPostId, 0, 0, 0);
-        $this->httpClient->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
-        $this->assert404NotFoundHeader();
+        $this->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
+        $this->assertResponseCode(404);
     }
 
     public function testBlogPostDetailOnlinePrivateLoggedOut()
     {
         $blogPostId = 2;
         $this->changeBlogPost($blogPostId, 1);
-        $this->httpClient->followOneRedirectForNextRequest();
-        $this->httpClient->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
-        $this->assertAccessDeniedWithRedirectToLoginForm();
+        $requestUrl = $this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel');
+        $this->get($requestUrl);
+        $this->assertRedirectContains($this->Slug->getLogin($requestUrl));
+        $this->assertAccessDeniedFlashMessage();
     }
 
     public function testBlogPostDetailOnlinePrivateLoggedIn()
@@ -55,15 +59,15 @@ class BlogPostsControllerTest extends AppCakeTestCase
         $this->loginAsCustomer();
         $blogPostId = 2;
         $this->changeBlogPost($blogPostId, 1);
-        $this->httpClient->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
-        $this->assert200OkHeader();
+        $this->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
+        $this->assertResponseCode(200);
     }
 
     public function testBlogPostDetaiNonExistingLoggedOut()
     {
         $blogPostId = 3;
-        $this->httpClient->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
-        $this->assert404NotFoundHeader();
+        $this->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
+        $this->assertResponseCode(404);
     }
 
     public function testBlogPostDetailOnlinePublicManufacturerPrivateLoggedOut()
@@ -72,9 +76,10 @@ class BlogPostsControllerTest extends AppCakeTestCase
         $manufacturerId = 15;
         $this->changeBlogPost($blogPostId, 0, $manufacturerId);
         $this->changeManufacturer($manufacturerId, 'is_private', 1);
-        $this->httpClient->followOneRedirectForNextRequest();
-        $this->httpClient->get($this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel'));
-        $this->assertAccessDeniedWithRedirectToLoginForm();
+        $requestUrl = $this->Slug->getBlogPostDetail($blogPostId, 'Demo Blog Artikel');
+        $this->get($requestUrl);
+        $this->assertRedirectContains($this->Slug->getLogin($requestUrl));
+        $this->assertAccessDeniedFlashMessage();
     }
 
     public function testBlogPostDetailChangeNullManufacturer()